fix: resolve test failure and suppress third-party warnings
- Fixed test_activation_shapes to expect float16 (Phase 18 FP16 optimization)
- Added pytest warning filters for PyTorch and datasets deprecation warnings
- All 207 unit tests now passing with zero warnings

Test Results:
- 207/207 tests passing (100%)
- 0 warnings
- Overall coverage: 43.50%

Changes:
- tests/unit/test_activation_service.py: Updated dtype assertion to float16
- pyproject.toml: Added filterwarnings to suppress third-party warnings

+++ backend/tests/unit/test_activation_service.py
@@ -456,7 +456,8 @@ class TestActivationService:
             # Shape should be [num_samples, seq_len, hidden_dim]
             # = [5, 10, 64]
             assert activations.shape == (5, 10, 64)
-            assert activations.dtype == np.float32
+            # System uses FP16 for memory efficiency (Phase 18 optimization)
+            assert activations.dtype == np.float16
 
     def test_different_architectures(self, activation_service, real_dataset_dir):
         """Test extraction works with different model architectures."""
